factoring out team creation to service
