+++ src/main/java/uk/ac/ebi/subs/api/controllers/TeamController.java
@@ -1,5 +1,7 @@
 package uk.ac.ebi.subs.api.controllers;
 
+import lombok.NonNull;
+import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.rest.webmvc.BasePathAwareController;
@@ -10,12 +12,18 @@ import org.springframework.security.access.method.P;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
+import uk.ac.ebi.subs.api.aap.TeamCreationService;
+import uk.ac.ebi.subs.api.aap.TeamDto;
+import uk.ac.ebi.subs.api.processors.TeamResourceProcessor;
 import uk.ac.ebi.subs.api.services.UserTeamService;
 import uk.ac.ebi.subs.data.component.Team;
 import uk.ac.ebi.subs.repository.repos.SubmissionRepository;
 import uk.ac.ebi.subs.repository.security.PreAuthorizeParamTeamName;
+import uk.ac.ebi.tsc.aap.client.model.User;
 
 import java.util.List;
 
@@ -24,19 +32,23 @@ import static org.springframework.hateoas.mvc.ControllerLinkBuilder.methodOn;
 
 @RestController
 @BasePathAwareController
+@RequiredArgsConstructor
 public class TeamController {
 
+    @NonNull
     private SubmissionRepository submissionRepository;
+
+    @NonNull
     private PagedResourcesAssembler<Team> teamPagedResourcesAssembler;
+
+    @NonNull
     private UserTeamService userTeamService;
 
-    public TeamController(SubmissionRepository submissionRepository,
-                          PagedResourcesAssembler<Team> teamPagedResourcesAssembler,
-                          UserTeamService userTeamService) {
-        this.submissionRepository = submissionRepository;
-        this.teamPagedResourcesAssembler = teamPagedResourcesAssembler;
-        this.userTeamService = userTeamService;
-    }
+    @NonNull
+    private TeamCreationService teamCreationService;
+
+    @NonNull
+    private TeamResourceProcessor teamResourceProcessor;
 
     @RequestMapping("/user/teams")
     public Resources<Resource<Team>> getTeams(Pageable pageable) {
@@ -67,6 +79,20 @@ public class TeamController {
         );
 
         return resource;
+    }
+
+    @RequestMapping(value = "/user/teams", method = RequestMethod.POST)
+    public Resource<Team> createTeam(@RequestBody TeamDto teamDto) {
+
+        final Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        User user = (User) authentication.getDetails();
+
+
+        Team team = teamCreationService.createTeam(user,teamDto);
+
+        Resource<Team> teamResource = new Resource<>(team);
+
+        return teamResourceProcessor.process(teamResource);
     }
