+++ src/main/resources/application.properties
@@ -8,9 +8,9 @@ spring.boot.admin.client.name=subs-api
 info.version=${version}
 info.name=${name}
 jwt.certificate=explore.der
-usi.aap.team_prefix=USI_
 aap.domains.url=https://explore.api.aai.ebi.ac.uk
 aap.enabled=true
 spring.main.banner-mode=off
 usi.tokenservice.username=${usiCoreAAPUsername}
 usi.tokenservice.password=${usiCoreAAPPassword}
+usi.teamName.prefix=subs.dev-team-
